perf: 字段标记icon情况下优化渲染
