+++ packages/s2-core/src/common/icons/gui-icon.ts
@@ -4,6 +4,7 @@
 import { Group, type ImageStyleProps } from '@antv/g';
 import { clone, omit } from 'lodash';
 import { CustomImage } from '../../engine';
+import { batchSetStyle } from '../../utils/g-utils';
 import { DebuggerUtil } from '../debug';
 import { getIcon } from './factory';
 
@@ -120,6 +121,17 @@ export class GuiIcon extends Group {
     });
 
     this.iconImageShape = image;
+    this.setImageAttrs({ name, fill });
+  }
+
+  public reRender(cfg: GuiIconCfg) {
+    this.name = cfg.name;
+    this.cfg = cfg;
+    const { name, fill } = this.cfg;
+    const attrs = clone(this.cfg);
+
+    this.iconImageShape.imgType = GuiIcon.type;
+    batchSetStyle(this.iconImageShape, omit(attrs, 'fill'));
     this.setImageAttrs({ name, fill });
   }
