+++ packages/s2-core/src/utils/g-renders.ts
@@ -9,7 +9,6 @@ import {
   Polygon,
   Polyline,
   Rect,
-  type BaseStyleProps,
   type CircleStyleProps,
   type DisplayObject,
   type LineStyleProps,
@@ -144,13 +143,4 @@ export function renderTreeIcon(options: {
   }
 
   return iconShape;
-}
-
-export function batchSetStyle<
-  T extends DisplayObject,
-  S extends BaseStyleProps,
->(obj: T, style: S) {
-  for (const styleKey in style) {
-    obj.style[styleKey] = style[styleKey];
-  }
 }
