+++ packages/s2-core/src/utils/g-utils.ts
@@ -1,0 +1,10 @@
+import { BaseStyleProps, DisplayObject } from '@antv/g';
+
+export function batchSetStyle<
+  T extends DisplayObject,
+  S extends BaseStyleProps,
+>(obj: T, style: S) {
+  for (const styleKey in style) {
+    obj.style[styleKey] = style[styleKey];
+  }
+}
